+++ scripts/update_test.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-echo "=== CARGO UPDATE ===" &&
-cargo update &&
-echo "=== TEST.SH ===" &&
-./test.sh